split manager in files per state
